added integration test for Backend Model and Backend Type
